+++ modules/gapi/include/opencv2/gapi/infer/ie.hpp
@@ -17,7 +17,7 @@
 
 #include <opencv2/core/cvdef.h>     // GAPI_EXPORTS
 #include <opencv2/gapi/gkernel.hpp> // GKernelPackage
-#include <opencv2/gapi/infer.hpp> // Generic
+#include <opencv2/gapi/infer.hpp>   // Generic
 
 namespace cv {
 namespace gapi {
